+++ package.json
@@ -11,7 +11,7 @@
     "dist": "npm run dist-main && npm run dist-renderer",
     "start": "cross-env NODE_ENV=production electron ./main",
     "start-hot": "cross-env HOT=1 NODE_ENV=development electron -r babel-register -r babel-polyfill ./main.dev",
-    "hot-server": "cross-env HOT=1 NODE_ENV=development node -r babel-register server.js",
+    "hot-server": "cross-env HOT=1 NODE_ENV=development node -r babel-register server.dev.js",
     "dev": "concurrently --kill-others \"npm run hot-server\" \"npm run start-hot\"",
     "release": "npm run dist && npm run package-all",
     "package-all": "build -mwl",
